auto-improve: [#799 Step 2/6] Split `cai_lib/cmd_helpers.py` into cohesive submodules

Refs damien-robotsix/robotsix-cai#819
